feat: replace levenshtein usage with jaro-winkler

+++ pkg/malcontent/malcontent.go
@@ -96,7 +96,7 @@ type FileReport struct {
 	// The relative path we think this moved from.
 	PreviousRelPath string `json:",omitempty" yaml:",omitempty"`
 	// The levenshtein distance between the previous path and the current path
-	PreviousRelPathScore float64 `json:",omitempty" yaml:",omitempty"`
+	PreviousRelPathScore float32 `json:",omitempty" yaml:",omitempty"`
 	PreviousRiskScore    int     `json:",omitempty" yaml:",omitempty"`
 	PreviousRiskLevel    string  `json:",omitempty" yaml:",omitempty"`
 
@@ -148,5 +148,5 @@ type CombinedReport struct {
 	AddedFR   *FileReport
 	Removed   string
 	RemovedFR *FileReport
-	Score     float64
+	Score     float32
 }
